Fix settings.py, requirements.txt
When I tried to run the project, python gave me two errors.  The first
was that the `=` operator wasn't allowed in the requirements.txt.  It
told me to change it to `==` which I did and it correctly installed the
dependencies.

The second error that I came upon was a syntax error when running the
project.  It said that `DIRS: []` wasn't valid and for Python, it's not.
This updates it to assign the `DIRS` variable the array like `DIRS=[]`